+++ CHANGES.md
@@ -1,5 +1,10 @@
 # parameters Changelog
 
+## 1.0.0 (2023-01-27)
+
+- BREAKING CHANGE: Dropping Python 2.7 support
+- Introduce new class PintParameterSet for solving formulas with units
+
 ## 0.7.1 (2023-01-25)
 
 - Fix tests: must set loc in all monte carlo runs